♻️ Refactor: 돈 관련 자료형을 Long 으로 변경

+++ backend/donation/src/main/java/kr/co/olivepay/donation/entity/Donation.java
@@ -24,10 +24,10 @@ public class Donation extends BaseEntity {
     @JoinColumn(name = "donor_id", nullable = false, columnDefinition = "INT UNSIGNED")
     private Donor donor;
 
-    private Integer money;
+    private Long money;
 
     @Builder
-    public Donation(Long franchiseId, Donor donor, Integer money) {
+    public Donation(Long franchiseId, Donor donor, Long money) {
         this.franchiseId = franchiseId;
         this.donor = donor;
         this.money = money;
